// src/lib/study-lenses--deprecated-architecture/lib/quizzing/context/types.ts
export type UsageKind = 'declared' | 'read' | 'assigned' | 'read-and-assigned';
// ...
export type IdentifierAnchor = Readonly<{
	range: readonly [number, number];
	name: string;
	usageKind: UsageKind;
}>;
// ...
export type PropertyAccessAnchor = Readonly<{
	range: readonly [number, number];
	name: string;
}>;